+++ src/errors.rs
@@ -28,7 +28,7 @@ impl std::fmt::Display for FxError {
             FxError::OpenNewWindow => {
                 "Error: Cannot open this type of item in new window".to_owned()
             }
-            FxError::TomlDe => "Error: Cannot deserialize toml".to_owned(),
+            FxError::TomlDe => "Error: Cannot deserialize toml: Check your config file.".to_owned(),
             FxError::TomlSer => "Error: Cannot serialize toml".to_owned(),
             FxError::WalkDir => "Error: Cannot read directory".to_owned(),
             FxError::Encode => "Error: Incorrect encoding".to_owned(),